Refactored Langflow code to lib/
